상점 물가 반영, 생산량, 판매액, 구입액 갱신 로직 추가

+++ Assets/Scripts/UI/Entity/UIItemIcon.cs
@@ -26,7 +26,7 @@ public class UIItemIcon : MonoBehaviour
         kCountText.text = "";
     }
 
-    public void SetPrice(int _price)
+    public void SetPrice(long _price)
     {
         kPriceText.text = _price.ToString() + " G";
     }
